Add PaginatedCollector with #filtered_set
Besides that it can pass a query hash, to the underlining client, it also receives a filter predicate as a form of Proc (or lambda).

This offers the functionality that instead of keeping all the results in memory and then filtering through, we do it per page thus decreasing memory need.