Make API fetching async with Concurrent::Future and Agent
The idea is that we download the first page on a blocking manner then we check
the response if we have more pages.
If we do, we create a Future for every page, collecting the result in a thread-safe
manner into an Agent.

After all the pages futures have been completed (mapping on the job
values), then we wait for all the agent update to finish. In the end, we
will have the Agent’s Set as the result.

I had to disable the tests for this version because neither webmock nor
minitest mocks are thread safe :(